# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  novel.ErrorResponse:
    properties:
      code:
        description: 错误码（非0表示错误）
        type: integer
      detail:
        description: 错误详情（可选）
        type: string
      message:
        description: 错误消息
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@lemon.ai
    name: API Support
  description: |-
    Lemon is an AI-powered API service built with Eino framework.
    It provides LLM chat, intelligent agent, and more AI capabilities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemon API
  version: "1.0"
paths:
...
  /api/v1/narrations/{narration_id}/audios:
    get:
      responses: {}
    post:
      consumes:
      - application/json
      description: 为章节解说生成所有章节音频片段，使用TTS服务生成音频。音频生成是异步的，提交任务后需要通过状态查询接口轮询进度。
      parameters:
      - description: 解说ID
        in: path
        name: narration_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '成功响应"  "{\"code\": 0, \"message\": \"音频生成任务已提交\", \"data\":
            {\"audio_ids\": [\"...\"], \"count\": 1, \"narration_id\": \"...\"}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
      summary: 生成章节音频
      tags:
      - 音频生成
  /api/v1/narrations/{narration_id}/audios/versions:
    get:
      consumes:
      - application/json
      description: 获取章节解说的所有音频版本号列表。
      parameters:
      - description: 解说ID
        in: path
        name: narration_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '成功响应"  "{\"code\": 0, \"message\": \"获取成功\", \"data\": {\"narration_id\":
            \"...\", \"versions\": [1, 2, 3]}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
      summary: 获取音频版本列表
      tags:
      - 音频生成
...
  /api/v1/novels/chapters/{chapter_id}/videos/narration:
    post:
      consumes:
      - application/json
      description: 为章节生成所有 narration 视频，所有分镜都单独生成视频，使用图生视频方式（Ark API 或 FFmpeg）。视频生成是异步的，提交任务后需要通过状态查询接口轮询进度。
      parameters:
      - description: 章节ID
        in: path
        name: chapter_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '成功响应"  "{\"code\": 0, \"message\": \"视频生成任务已提交\", \"data\":
            {\"video_ids\": [\"...\"], \"count\": 1, \"chapter_id\": \"...\"}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/novel.ErrorResponse'
      summary: 生成章节的 narration 视频
      tags:
      - 视频生成
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"